Allow resources to auth with multiple methods
This change allows resources (e.g. User) to authenticate through
multiple methods for the same real life person. Prior to this, each auth
method (e.g. facebook, twitter, etc) would create a new row in User when
it was used.

See notes in the README that have changed for this commit for more
information.